+++ 04_GlobalVariables.sol
@@ -2,10 +2,18 @@
 pragma solidity ^0.8.7;
 
 contract GlobalVariables {
-    function globalVars() external view returns (address, uint, uint) {
+    function globalVars()
+        external
+        view
+        returns (
+            address,
+            uint256,
+            uint256
+        )
+    {
         address sender = msg.sender;
-        uint timestamp = block.timestamp;
-        uint blockNum = block.number;
+        uint256 timestamp = block.timestamp;
+        uint256 blockNum = block.number;
         return (sender, timestamp, blockNum);
     }
 }
