+++ 05_DefaultValues.sol
@@ -3,8 +3,8 @@ pragma solidity ^0.8.7;
 
 contract DefaultValues {
     bool public b; // false
-    uint public u; // 0
-    int public i; // 0
+    uint256 public u; // 0
+    int256 public i; // 0
     address public a; // 0x0000000000000000000000000000000000000000
     bytes32 public b32; // 0x0000000000000000000000000000000000000000000000000000000000000000
     // mapping, structs, enums, fixed size arrays
