+++ 07_Func.sol
@@ -3,11 +3,11 @@ pragma solidity ^0.8.7;
 
 contract FunctionIntro {
     // pure is view only.
-    function add(uint x, uint y) external pure returns(uint) {
+    function add(uint256 x, uint256 y) external pure returns (uint256) {
         return x + y;
     }
 
-    function sub(uint x, uint y) external pure returns(uint) {
+    function sub(uint256 x, uint256 y) external pure returns (uint256) {
         return x - y;
     }
 }
