+++ 09_Counter.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.7;
 
 contract Counter {
-    uint public count;
+    uint256 public count;
 
     function inc() external {
         count += 1;
@@ -11,5 +11,4 @@ contract Counter {
     function dec() external {
         count -= 1;
     }
-
 }
