+++ 10_Constructor.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.8.7;
 
 contract Constructor {
     address public owner;
-    uint public x;
-    
-    constructor(uint _x_1) {
+    uint256 public x;
+
+    constructor(uint256 _x_1) {
         owner = msg.sender;
         x = _x_1;
     }
