+++ 16_IterableMapping.sol
@@ -2,32 +2,32 @@
 pragma solidity ^0.8.7;
 
 contract IterableMapping {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
     mapping(address => bool) public inserted;
     address[] public keys;
 
-    function set(address _key, uint _val) external {
+    function set(address _key, uint256 _val) external {
         balances[_key] = _val;
 
-        if(!inserted[_key]) {
+        if (!inserted[_key]) {
             inserted[_key] = true;
             keys.push(_key);
         }
     }
 
-    function getSize() external view returns(uint) {
+    function getSize() external view returns (uint256) {
         return keys.length;
     }
 
-    function first() external view returns (uint) {
+    function first() external view returns (uint256) {
         return balances[keys[0]];
     }
 
-    function last() external view returns (uint) {
+    function last() external view returns (uint256) {
         return balances[keys[keys.length - 1]];
     }
 
-    function get(uint _i) external view returns (uint) {
+    function get(uint256 _i) external view returns (uint256) {
         return balances[keys[_i]];
     }
 }
