+++ 20_Error.sol
@@ -6,18 +6,19 @@ pragma solidity ^0.8.7;
 // custom error - save gas
 
 contract Error {
-    function testRequire(uint _i) public pure {
+    function testRequire(uint256 _i) public pure {
         require(_i <= 10, "i > 10");
         //code
     }
 
     // same function but better
-    function testRevert(uint _i) public pure {
-        if(_i > 10) {
+    function testRevert(uint256 _i) public pure {
+        if (_i > 10) {
             revert("i > 10");
         }
-}
-    uint public num = 123;
+    }
+
+    uint256 public num = 123;
 
     function testAssert() public view {
         assert(num == 123);
@@ -28,10 +29,10 @@ contract Error {
         num += 1;
     }
 
-    error myError(address caller, uint i);
+    error myError(address caller, uint256 i);
 
-    function testCustomError(uint _i) public view {
-        if(_i > 10) {
+    function testCustomError(uint256 _i) public view {
+        if (_i > 10) {
             revert myError(msg.sender, _i);
         }
     }
